requirements.txt
yt-dlp
ffmpeg-python
tqdm